requirements.txt
matplotlib==3.10.8
numpy==2.4.4
pandas==3.0.2
Pillow==12.2.0
seaborn==0.13.2
streamlit==1.55.0
